manus/requirements.txt
fastapi
uvicorn[standard]
gunicorn
line-bot-sdk
httpx
python-dotenv
sqlalchemy[asyncio]
aiosqlite
jinja2
python-multipart